requirements.txt
asn1crypto==0.24.0
astroid==2.0.4
backcall==0.1.0
beautifulsoup4==4.6.3
bleach==3.0.2
bs4==0.0.1
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
colorama==0.4.0
cryptography==2.1.4
decorator==4.3.0
defusedxml==0.5.0
distro==1.3.0
Django==2.1.3
django-background-task==0.1.8
django-background-tasks==1.2.0
django-compat==1.0.15
django-pdfkit==0.3.1
django-postgresql==0.0.3
django-rest-framework==0.1.0
djangorestframework==3.9.0
entrypoints==0.2.3
et-xmlfile==1.0.1
httplib2==0.9.2
idna==2.6
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
jdcal==1.4
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
keyring==10.6.0
keyrings.alt==3.0
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.4
mysql-connector==2.1.7
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.2
numpy==1.15.4
oauth==1.0.1
oauthlib==2.1.0
olefile==0.45.1
openpyxl==2.5.9
panda==0.3.1
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pbr==5.1.0
pdfkit==0.6.0
pexpect==4.2.1
pickleshare==0.7.5
Pillow==5.1.0
prometheus-client==0.4.2
prompt-toolkit==2.0.7
protobuf==3.0.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.2.0
PyJWT==1.6.4
pylint==2.1.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-dateutil==2.7.5
python-debian==0.1.32
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.7
pywin32-ctypes==0.2.0
pywinpty==0.5.4
pyxdg==0.25
PyYAML==3.12
pyzmq==17.1.2
qtconsole==4.4.3
reportlab==3.4.0
requests==2.18.4
requests-oauthlib==1.0.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.13.2
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
stevedore==1.30.0
terminado==0.8.1
testpath==0.4.2
testresources==2.0.1
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.1.0
urllib3==1.24.1
virtualenv==16.1.0
virtualenv-clone==0.4.0
virtualenvwrapper==4.8.2
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
win-unicode-console==0.5
wrapt==1.10.11
xlrd==1.1.0
zope.interface==4.3.2
zxcvbn==4.4.27
